+++ src/features/whatsnew/whatsNew.js
@@ -23,11 +23,19 @@ const MESSAGE = `[<img width="130" alt="get in touch with Consensys Diligence" s
 
 
 
-Hey there 🙌 **Solidity Visual Auditor** just got better! Find out more below.
+Hey there 🙌 **Solidity Visual Developer** just got better! Find out more below.
 
 ### What's New?
 
 The complete changelog can be found [here](https://github.com/ConsenSys/vscode-solidity-auditor/blob/master/CHANGELOG.md). 
+
+#### v0.0.27
+- new: \`Solidity Visual Auditor\` is now \`Solidity Visual Developer\` 🎉
+- new: Ethereum Address hover commands. Hover over an ethereum account address to:
+    - \`open\` the account on etherscan.io
+    - show the contract \`code\`
+    - show the \`VerifiedContract\` source code
+    - \`decompile\` the byte-code. requires [vscode-decompiler](https://marketplace.visualstudio.com/items?itemName=tintinweb.vscode-decompiler)
 
 #### v0.0.26
 - new: support for solidity \`>= 0.6.0\`